Move `defaultConfig` logic to its own file
